+++ src/lib/legacy.ts
@@ -796,6 +796,35 @@ const dataPoints = {
     wooxSwitch: 102,
     wooxBattery: 14,
     wooxSmokeTest: 8,
+
+    // Woox thermostat
+    wooxDormancy: 108, // ???
+    wooxRefresh: 120, //  ???
+    wooxControlTemperature: 119, // map auto and manual temperature setpoint.
+    wooxManualTemperatureSetpoint: 16, //RW
+    wooxAutomaticTemperatureSetpoint: 105, //RW
+    wooxMode: 2, //RW
+    wooxLocalTemperature: 24, //R
+    wooxTemperatureCalibration: 104, //RW
+    wooxWindowStatus: 107, //R open,close
+    wooxWindowTemperature: 116, //RW
+    wooxWindowTime: 117, //RW
+    wooxChildLock: 30, //RW
+    wooxBatteryCapacity: 34, //R
+    wooxEnergySavingTemperature: 102, //RW
+    wooxComfortTemperature: 101, //RW
+    wooxHolidayModeSettings: 103, //RW
+    wooxProgrammingMonday: 109, //RW
+    wooxProgrammingTuesday: 110, //RW
+    wooxProgrammingWednesday: 111, //RW
+    wooxProgrammingThursday: 112, //RW
+    wooxProgrammingFriday: 113, //RW
+    wooxProgrammingSaturday: 114, //RW
+    wooxProgrammingSunday: 115, //RW
+    wooxBoostHeating: 106, //RW
+    wooxFaultAlarm: 45, // R
+    wooxBoostHeatingCountdown: 118, //R
+
     // FrankEver
     frankEverTimer: 9,
     frankEverTreshold: 101,
@@ -1352,6 +1381,107 @@ const fromZigbee = {
             }
         },
     } satisfies Fz.Converter,
+
+    woox_thermostat: {
+        cluster: "manuSpecificTuya",
+        type: ["commandDataResponse", "commandDataReport"],
+        convert: (model, msg, publish, options, meta) => {
+            const result: KeyValueAny = {};
+            for (const dpValue of msg.data.dpValues) {
+                const dp = dpValue.dp;
+                const value = getDataValue(dpValue);
+                switch (dp) {
+                    case dataPoints.wooxMode:
+                        if (value === 0) {
+                            result.system_mode = "auto";
+                            result.away_mode = "OFF";
+                        } else if (value === 1) {
+                            result.system_mode = "heat";
+                            result.away_mode = "OFF";
+                        } else if (value === 2) {
+                            result.away_mode = "ON";
+                            result.system_mode = "auto";
+                        } else {
+                            result.away_mode = "OFF";
+                            result.system_mode = "off";
+                        }
+                        break;
+                    case dataPoints.wooxManualTemperatureSetpoint:
+                        result.current_heating_setpoint = Number.parseFloat((value / 2).toFixed(1));
+                        result.manual_heating_setpoint = Number.parseFloat((value / 2).toFixed(1));
+                        break;
+                    case dataPoints.wooxAutomaticTemperatureSetpoint:
+                        result.current_heating_setpoint = Number.parseFloat((value / 2).toFixed(1));
+                        result.auto_heating_setpoint = Number.parseFloat((value / 2).toFixed(1));
+                        break;
+                    case dataPoints.wooxLocalTemperature:
+                        result.local_temperature = Number.parseFloat((value / 10).toFixed(1));
+                        break;
+                    case dataPoints.wooxTemperatureCalibration:
+                        result.local_temperature_calibration = Number.parseFloat((value / 10).toFixed(1));
+                        break;
+                    case dataPoints.wooxWindowStatus:
+                        result.window_detection = value[0] ? "OPEN" : "CLOSED";
+                        break;
+                    case dataPoints.wooxWindowTemperature:
+                        result.window_detection_temperature = Number.parseFloat((value / 2).toFixed(1));
+                        break;
+                    case dataPoints.wooxWindowTime:
+                        result.window_detection_time = value;
+                        break;
+                    case dataPoints.wooxChildLock:
+                        result.child_lock = value ? "LOCK" : "UNLOCK";
+                        break;
+                    case dataPoints.wooxBatteryCapacity:
+                        result.battery = value;
+                        result.battery_low = value < 30 ? 1 : 0;
+                        break;
+                    case dataPoints.wooxBoostHeatingCountdown:
+                        result.boost_time = value;
+                        break;
+                    case dataPoints.wooxEnergySavingTemperature:
+                        result.eco_temperature = Number.parseFloat((value / 2).toFixed(1));
+                        break;
+                    case dataPoints.wooxComfortTemperature:
+                        result.comfort_temperature = Number.parseFloat((value / 2).toFixed(1));
+                        break;
+                    case dataPoints.wooxBoostHeating:
+                        result.boost_heating = value ? "ON" : "OFF";
+                        break;
+                    case dataPoints.wooxFaultAlarm:
+                        result.error_status = value;
+                        break;
+                    case dataPoints.wooxProgrammingMonday:
+                        result.monday_schedule = value.join();
+                        break;
+                    case dataPoints.wooxProgrammingTuesday:
+                        result.tuesday_schedule = value.join();
+                        break;
+                    case dataPoints.wooxProgrammingWednesday:
+                        result.wednesday_schedule = value.join();
+                        break;
+                    case dataPoints.wooxProgrammingThursday:
+                        result.thursday_schedule = value.join();
+                        break;
+                    case dataPoints.wooxProgrammingFriday:
+                        result.friday_schedule = value.join();
+                        break;
+                    case dataPoints.wooxProgrammingSaturday:
+                        result.saturday_schedule = value.join();
+                        break;
+                    case dataPoints.wooxProgrammingSunday:
+                        result.sunday_schedule = value.join();
+                        break;
+                    case dataPoints.wooxHolidayModeSettings:
+                        result.holidays_schedule = value.join();
+                        break;
+                }
+                logger.debug(`Unrecognized DP #${dp} with data ${JSON.stringify(dpValue.data)}`, "zhc:legacy:fz:woox_thermostat");
+            }
+            return result;
+        },
+    } satisfies Fz.Converter,
+
     hpsz: {
         cluster: "manuSpecificTuya",
         type: ["commandDataResponse", "commandDataReport"],
@@ -6764,6 +6894,162 @@ const toZigbee2 = {
             } else {
                 throw new Error(`Not supported: '${key}'`);
             }
+        },
+    } satisfies Tz.Converter,
+    woox_thermostat_child_lock: {
+        key: ["child_lock"],
+        convertSet: async (entity, key, value, meta) => {
+            await sendDataPointBool(entity, dataPoints.wooxChildLock, value === "LOCK");
+        },
+    } satisfies Tz.Converter,
+    woox_thermostat_current_heating_setpoint: {
+        key: ["current_heating_setpoint"],
+        // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+        convertSet: async (entity, key, value: any, meta) => {
+            const temp = Math.round(value * 10);
+            await sendDataPointValue(entity, dataPoints.wooxControlTemperature, temp);
+        },
+    } satisfies Tz.Converter,
+    woox_thermostat_system_mode: {
+        key: ["system_mode"],
+        convertSet: async (entity, key, value, meta) => {
+            if (value === "auto") {
+                await sendDataPointEnum(entity, dataPoints.wooxMode, 0);
+                await sendDataPointValue(entity, dataPoints.wooxControlTemperature, 220);
+                return {state: {current_heating_setpoint: 22}};
+            }
+
+            if (value === "heat") {
+                await sendDataPointEnum(entity, dataPoints.wooxMode, 1);
+                await sendDataPointValue(entity, dataPoints.wooxControlTemperature, 170);
+                return {state: {current_heating_setpoint: 17}};
+            }
+
+            logger.debug(`Woox thermostat: uknown mode ${value}`, "woox_thermostat_system_mode");
+        },
+    } satisfies Tz.Converter,
+
+    woox_away_mode: {
+        key: ["away_mode"],
+        convertSet: async (entity, key, value, meta) => {
+            if (value === "ON") {
+                await sendDataPointEnum(entity, dataPoints.wooxMode, 2);
+                return {state: {current_heating_setpoint: 0}};
+            }
+
+            await sendDataPointEnum(entity, dataPoints.wooxMode, 0);
+            await sendDataPointValue(entity, dataPoints.wooxControlTemperature, 220);
+            return {state: {current_heating_setpoint: 22}};
+        },
+    } satisfies Tz.Converter,
+
+    woox_comfort_temperature: {
+        key: ["comfort_temperature"],
+        // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+        convertSet: async (entity, key, value: any, meta) => {
+            const temp = Math.round(value * 2);
+            await sendDataPointValue(entity, dataPoints.wooxComfortTemperature, temp);
+        },
+    } satisfies Tz.Converter,
+
+    woox_eco_temperature: {
+        key: ["eco_temperature"],
+        // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+        convertSet: async (entity, key, value: any, meta) => {
+            const temp = Math.round(value * 2);
+            await sendDataPointValue(entity, dataPoints.wooxEnergySavingTemperature, temp);
+        },
+    } satisfies Tz.Converter,
+
+    woox_local_temperature_calibration: {
+        key: ["local_temperature_calibration"],
+        // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+        convertSet: async (entity, key, value: any, meta) => {
+            let val = value;
+            if (val < 0) {
+                val = val + 4096;
+            }
+            const temp = Math.round(val * 10);
+            await sendDataPointValue(entity, dataPoints.wooxTemperatureCalibration, temp);
+        },
+    } satisfies Tz.Converter,
+
+    woox_window_detection_temperature: {
+        key: ["window_detection_temperature"],
+        // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+        convertSet: async (entity, key, value: any, meta) => {
+            const temp = Math.round(value * 2);
+            await sendDataPointValue(entity, dataPoints.wooxWindowTemperature, temp);
+        },
+    } satisfies Tz.Converter,
+
+    woox_window_detection_time: {
+        key: ["window_detection_time"],
+        convertSet: async (entity, key, value, meta) => {
+            await sendDataPointValue(entity, dataPoints.wooxWindowTime, value);
+        },
+    } satisfies Tz.Converter,
+
+    woox_boost_heating: {
+        key: ["boost_heating"],
+        convertSet: async (entity, key, value, meta) => {
+            await sendDataPointBool(entity, dataPoints.wooxBoostHeating, value === "ON");
+        },
+    } satisfies Tz.Converter,
+
+    woox_holidays_schedule: {
+        key: ["holidays_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("holidays_schedule", "woox_holidays_schedule");
+        },
+    } satisfies Tz.Converter,
+
+    woox_monday_schedule: {
+        key: ["monday_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("monday_schedule", "woox_monday_schedule");
+        },
+    } satisfies Tz.Converter,
+
+    woox_tuesday_schedule: {
+        key: ["tuesday_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("tuesday_schedule", "woox_tuesday_schedule");
+        },
+    } satisfies Tz.Converter,
+
+    woox_wednesday_schedule: {
+        key: ["wednesday_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("wednesday_schedule", "woox_wednesday_schedule");
+        },
+    } satisfies Tz.Converter,
+
+    woox_thursday_schedule: {
+        key: ["thursday_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("thursday_schedule", "woox_thursday_schedule");
+        },
+    } satisfies Tz.Converter,
+
+    woox_friday_schedule: {
+        key: ["friday_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("friday_schedule", "woox_friday_schedule");
+        },
+    } satisfies Tz.Converter,
+
+    woox_saturday_schedule: {
+        key: ["saturday_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("saturday_schedule", "woox_saturday_schedule");
+        },
+    } satisfies Tz.Converter,
+
+    woox_sunday_schedule: {
+        key: ["sunday_schedule"],
+        convertSet: async (entity, key, value, meta) => {
+            await logger.warning("sunday_schedule", "woox_sunday_schedule");
         },
     } satisfies Tz.Converter,
 };